hackthon - Copy/ml_model/requirements.txt
flask
flask-cors
scikit-learn
spacy
PyPDF2
pdfplumber
numpy
pandas
werkzeug
gunicorn
python-docx
